# Activate Automatic Mixed Precision.