--- a/mrcnn/__init__.py
+++ b/mrcnn/__init__.py
@@ -1,1 +1,3 @@
+import os
 
+os.environ['TF_ENABLE_AUTO_MIXED_PRECISION'] = '1'