Should increase performance on NVIDIA GPUs featuring TensorCores by a factor up to 3X (according to https://developer.nvidia.com/automatic-mixed-precision).