@maxfrei750 2019-05-17
@khoa-ho I did not have the time to benchmark it in depth. I did now, and unfortunately, it does not seem to make to much of a difference:

Without automatic mixed precision:
```
Epoch 1/10
100/100 [==============================] - 528s 5s/step - loss: 1.3636 - rpn_class_loss: 0.0515 - rpn_bbox_loss: 0.2642 - mrcnn_class_loss: 0.3703 - mrcnn_bbox_loss: 0.3133 - mrcnn_mask_loss: 0.3643 - val_loss: 0.9550 - val_rpn_class_loss: 0.0225 - val_rpn_bbox_loss: 0.2284 - val_mrcnn_class_loss: 0.2495 - val_mrcnn_bbox_loss: 0.2015 - val_mrcnn_mask_loss: 0.2530
Epoch 2/10
100/100 [==============================] - 135s 1s/step - loss: 0.8770 - rpn_class_loss: 0.0263 - rpn_bbox_loss: 0.1979 - mrcnn_class_loss: 0.2470 - mrcnn_bbox_loss: 0.1591 - mrcnn_mask_loss: 0.2468 - val_loss: 0.8652 - val_rpn_class_loss: 0.0202 - val_rpn_bbox_loss: 0.2140 - val_mrcnn_class_loss: 0.2527 - val_mrcnn_bbox_loss: 0.1438 - val_mrcnn_mask_loss: 0.2345
Epoch 3/10
100/100 [==============================] - 131s 1s/step - loss: 0.7819 - rpn_class_loss: 0.0221 - rpn_bbox_loss: 0.1797 - mrcnn_class_loss: 0.2229 - mrcnn_bbox_loss: 0.1288 - mrcnn_mask_loss: 0.2282 - val_loss: 0.8323 - val_rpn_class_loss: 0.0224 - val_rpn_bbox_loss: 0.1948 - val_mrcnn_class_loss: 0.2238 - val_mrcnn_bbox_loss: 0.1510 - val_mrcnn_mask_loss: 0.2402
Epoch 4/10
100/100 [==============================] - 134s 1s/step - loss: 0.7084 - rpn_class_loss: 0.0180 - rpn_bbox_loss: 0.1594 - mrcnn_class_loss: 0.2001 - mrcnn_bbox_loss: 0.1136 - mrcnn_mask_loss: 0.2173 - val_loss: 0.7762 - val_rpn_class_loss: 0.0216 - val_rpn_bbox_loss: 0.1889 - val_mrcnn_class_loss: 0.2172 - val_mrcnn_bbox_loss: 0.1300 - val_mrcnn_mask_loss: 0.2184
Epoch 5/10
100/100 [==============================] - 129s 1s/step - loss: 0.6606 - rpn_class_loss: 0.0151 - rpn_bbox_loss: 0.1381 - mrcnn_class_loss: 0.1893 - mrcnn_bbox_loss: 0.1068 - mrcnn_mask_loss: 0.2113 - val_loss: 0.7730 - val_rpn_class_loss: 0.0223 - val_rpn_bbox_loss: 0.1899 - val_mrcnn_class_loss: 0.2175 - val_mrcnn_bbox_loss: 0.1220 - val_mrcnn_mask_loss: 0.2213
Epoch 6/10
100/100 [==============================] - 130s 1s/step - loss: 0.6302 - rpn_class_loss: 0.0166 - rpn_bbox_loss: 0.1320 - mrcnn_class_loss: 0.1817 - mrcnn_bbox_loss: 0.0974 - mrcnn_mask_loss: 0.2025 - val_loss: 0.8015 - val_rpn_class_loss: 0.0286 - val_rpn_bbox_loss: 0.2213 - val_mrcnn_class_loss: 0.2165 - val_mrcnn_bbox_loss: 0.1140 - val_mrcnn_mask_loss: 0.2212
Epoch 7/10
100/100 [==============================] - 127s 1s/step - loss: 0.6206 - rpn_class_loss: 0.0165 - rpn_bbox_loss: 0.1315 - mrcnn_class_loss: 0.1780 - mrcnn_bbox_loss: 0.0926 - mrcnn_mask_loss: 0.2020 - val_loss: 0.7603 - val_rpn_class_loss: 0.0252 - val_rpn_bbox_loss: 0.1950 - val_mrcnn_class_loss: 0.2123 - val_mrcnn_bbox_loss: 0.1138 - val_mrcnn_mask_loss: 0.2140
Epoch 8/10
100/100 [==============================] - 130s 1s/step - loss: 0.5603 - rpn_class_loss: 0.0138 - rpn_bbox_loss: 0.1100 - mrcnn_class_loss: 0.1591 - mrcnn_bbox_loss: 0.0861 - mrcnn_mask_loss: 0.1912 - val_loss: 0.7732 - val_rpn_class_loss: 0.0230 - val_rpn_bbox_loss: 0.2363 - val_mrcnn_class_loss: 0.1903 - val_mrcnn_bbox_loss: 0.1107 - val_mrcnn_mask_loss: 0.2129
Epoch 9/10
100/100 [==============================] - 129s 1s/step - loss: 0.5300 - rpn_class_loss: 0.0112 - rpn_bbox_loss: 0.1020 - mrcnn_class_loss: 0.1531 - mrcnn_bbox_loss: 0.0775 - mrcnn_mask_loss: 0.1863 - val_loss: 0.7769 - val_rpn_class_loss: 0.0207 - val_rpn_bbox_loss: 0.1908 - val_mrcnn_class_loss: 0.2210 - val_mrcnn_bbox_loss: 0.1221 - val_mrcnn_mask_loss: 0.2223
Epoch 10/10
100/100 [==============================] - 131s 1s/step - loss: 0.5324 - rpn_class_loss: 0.0124 - rpn_bbox_loss: 0.1054 - mrcnn_class_loss: 0.1557 - mrcnn_bbox_loss: 0.0755 - mrcnn_mask_loss: 0.1834 - val_loss: 0.6683 - val_rpn_class_loss: 0.0166 - val_rpn_bbox_loss: 0.1756 - val_mrcnn_class_loss: 0.1667 - val_mrcnn_bbox_loss: 0.1058 - val_mrcnn_mask_loss: 0.2036
```

With automatic mixed precision:
```
Epoch 1/10
100/100 [==============================] - 493s 5s/step - loss: 1.2676 - rpn_class_loss: 0.0522 - rpn_bbox_loss: 0.2450 - mrcnn_class_loss: 0.3774 - mrcnn_bbox_loss: 0.3005 - mrcnn_mask_loss: 0.2925 - val_loss: 0.8634 - val_rpn_class_loss: 0.0219 - val_rpn_bbox_loss: 0.2071 - val_mrcnn_class_loss: 0.2417 - val_mrcnn_bbox_loss: 0.1535 - val_mrcnn_mask_loss: 0.2393
Epoch 2/10
100/100 [==============================] - 140s 1s/step - loss: 0.8763 - rpn_class_loss: 0.0254 - rpn_bbox_loss: 0.2031 - mrcnn_class_loss: 0.2473 - mrcnn_bbox_loss: 0.1562 - mrcnn_mask_loss: 0.2444 - val_loss: 0.8999 - val_rpn_class_loss: 0.0224 - val_rpn_bbox_loss: 0.2554 - val_mrcnn_class_loss: 0.2227 - val_mrcnn_bbox_loss: 0.1565 - val_mrcnn_mask_loss: 0.2429
Epoch 3/10
100/100 [==============================] - 137s 1s/step - loss: 0.7675 - rpn_class_loss: 0.0202 - rpn_bbox_loss: 0.1690 - mrcnn_class_loss: 0.2175 - mrcnn_bbox_loss: 0.1305 - mrcnn_mask_loss: 0.2304 - val_loss: 0.8560 - val_rpn_class_loss: 0.0204 - val_rpn_bbox_loss: 0.2361 - val_mrcnn_class_loss: 0.2233 - val_mrcnn_bbox_loss: 0.1428 - val_mrcnn_mask_loss: 0.2334
Epoch 4/10
100/100 [==============================] - 141s 1s/step - loss: 0.7257 - rpn_class_loss: 0.0207 - rpn_bbox_loss: 0.1590 - mrcnn_class_loss: 0.2077 - mrcnn_bbox_loss: 0.1194 - mrcnn_mask_loss: 0.2189 - val_loss: 0.7749 - val_rpn_class_loss: 0.0212 - val_rpn_bbox_loss: 0.1948 - val_mrcnn_class_loss: 0.2150 - val_mrcnn_bbox_loss: 0.1248 - val_mrcnn_mask_loss: 0.2191
Epoch 5/10
100/100 [==============================] - 137s 1s/step - loss: 0.6721 - rpn_class_loss: 0.0160 - rpn_bbox_loss: 0.1474 - mrcnn_class_loss: 0.1915 - mrcnn_bbox_loss: 0.1066 - mrcnn_mask_loss: 0.2106 - val_loss: 0.7963 - val_rpn_class_loss: 0.0194 - val_rpn_bbox_loss: 0.2412 - val_mrcnn_class_loss: 0.2033 - val_mrcnn_bbox_loss: 0.1175 - val_mrcnn_mask_loss: 0.2148
Epoch 6/10
100/100 [==============================] - 136s 1s/step - loss: 0.6457 - rpn_class_loss: 0.0154 - rpn_bbox_loss: 0.1368 - mrcnn_class_loss: 0.1803 - mrcnn_bbox_loss: 0.1027 - mrcnn_mask_loss: 0.2106 - val_loss: 0.8194 - val_rpn_class_loss: 0.0259 - val_rpn_bbox_loss: 0.1933 - val_mrcnn_class_loss: 0.2365 - val_mrcnn_bbox_loss: 0.1318 - val_mrcnn_mask_loss: 0.2320
Epoch 7/10
100/100 [==============================] - 135s 1s/step - loss: 0.5680 - rpn_class_loss: 0.0131 - rpn_bbox_loss: 0.1161 - mrcnn_class_loss: 0.1639 - mrcnn_bbox_loss: 0.0846 - mrcnn_mask_loss: 0.1902 - val_loss: 0.7467 - val_rpn_class_loss: 0.0199 - val_rpn_bbox_loss: 0.2069 - val_mrcnn_class_loss: 0.2069 - val_mrcnn_bbox_loss: 0.1032 - val_mrcnn_mask_loss: 0.2099
Epoch 8/10
100/100 [==============================] - 133s 1s/step - loss: 0.5963 - rpn_class_loss: 0.0145 - rpn_bbox_loss: 0.1223 - mrcnn_class_loss: 0.1773 - mrcnn_bbox_loss: 0.0881 - mrcnn_mask_loss: 0.1941 - val_loss: 0.7582 - val_rpn_class_loss: 0.0249 - val_rpn_bbox_loss: 0.2062 - val_mrcnn_class_loss: 0.1996 - val_mrcnn_bbox_loss: 0.1171 - val_mrcnn_mask_loss: 0.2103
Epoch 9/10
100/100 [==============================] - 135s 1s/step - loss: 0.5570 - rpn_class_loss: 0.0147 - rpn_bbox_loss: 0.1148 - mrcnn_class_loss: 0.1563 - mrcnn_bbox_loss: 0.0817 - mrcnn_mask_loss: 0.1895 - val_loss: 0.7197 - val_rpn_class_loss: 0.0230 - val_rpn_bbox_loss: 0.2122 - val_mrcnn_class_loss: 0.1829 - val_mrcnn_bbox_loss: 0.1026 - val_mrcnn_mask_loss: 0.1990
Epoch 10/10
100/100 [==============================] - 133s 1s/step - loss: 0.5249 - rpn_class_loss: 0.0102 - rpn_bbox_loss: 0.1033 - mrcnn_class_loss: 0.1500 - mrcnn_bbox_loss: 0.0766 - mrcnn_mask_loss: 0.1847 - val_loss: 0.7504 - val_rpn_class_loss: 0.0346 - val_rpn_bbox_loss: 0.2366 - val_mrcnn_class_loss: 0.1848 - val_mrcnn_bbox_loss: 0.0968 - val_mrcnn_mask_loss: 0.1976
```

However, I still think that it might make sense to merge this PR. It does not impede the performance and might be beneficial at some point in the future.